Upgrade XGS saibcm-modules to 8.4 (#16246)
Why I did it
XGS saibcm-modules 8.4 is needed. #14471

Work item tracking
Microsoft ADO (number only): 24917414
How I did it
Copy files from xgs SDK 8.4 repo and modify makefiles to build the image.
Upgrade version to 8.4.0.2 in saibcm-modules.mk.

How to verify it
Build a private image and run full qualification with it: https://elastictest.org/scheduler/testplan/650419cb71f60aa92c456a2b

+++ platform/broadcom/saibcm-modules/systems/linux/kernel/modules/knet-cb/psample-cb.h
@@ -42,7 +42,7 @@ typedef struct {
     struct list_head list;
     struct net_device *dev;
     uint16 id;
-    uint8  port;
+    uint16 port;
     uint16 vlan;
     uint16 qnum;
     uint32 sample_rate;
